+++ src/enums/transferHandlerErrorEnum.js
@@ -1,0 +1,3 @@
+exports.INVALID_REQUEST = 'invalid request'
+exports.AMOUNT_MUST_BE_POSITIVE = 'amount must be positive'
+exports.INSUFFICIENT_XP = 'insufficient xp'