+++ test/indexTest.js
@@ -1,8 +1,16 @@
-const queryHandler = require('../src/handlers/queryHandler.js')
-const index = require('../src/index.js')
 const sinon = require('sinon')
 const expect = require('chai').expect
-const queryErrorHandler = require('../src/handlers/queryErrorHandler.js');
+const index = require('../src/index.js')
+const queryHandler = require('../src/handlers/queryHandler.js')
+const queryErrorHandler = require('../src/handlers/queryErrorHandler.js')
+const depositHandler = require('../src/handlers/depositHandler.js')
+const depositErrorHandler = require('../src/handlers/depositErrorHandler.js')
+const balanceHandler = require('../src/handlers/balanceHandler.js')
+const balanceErrorHandler = require('../src/handlers/balanceErrorHandler.js')
+const withdrawHandler = require('../src/handlers/withdrawHandler.js')
+const withdrawErrorHandler = require('../src/handlers/withdrawErrorHandler.js')
+const transferHandler = require('../src/handlers/transferHandler.js')
+const transferErrorHandler = require('../src/handlers/transferErrorHandler.js')
 
 const EVENT_QUERY = {
     requestContext: {
@@ -11,6 +19,36 @@ const EVENT_QUERY = {
     queryStringParameters: {
         userId: 'player'
     }
+}
+
+const EVENT_DEPOSIT = {
+    requestContext: {
+        routeKey: 'POST /xp/deposit'
+    },
+    body: '{"userId": "player", "amount": "100"}'
+}
+
+const EVENT_BALANCE = {
+    requestContext: {
+        routeKey: 'GET /xp/balance'
+    },
+    queryStringParameters: {
+        userId: 'player'
+    }
+}
+
+const EVENT_WITHDRAW = {
+    requestContext: {
+        routeKey: 'POST /xp/withdraw'
+    },
+    body: '{"userId": "player", "amount": "100"}'
+}
+
+const EVENT_TRANSFER = {
+    requestContext: {
+        routeKey: 'POST /xp/transfer'
+    },
+    body: '{"payerUserId": "player", "payeeUserId": "player2", "amount": "100"}'
 }
 
 const EVENT_INVALID_PATH = {
@@ -24,32 +62,138 @@ const QUERY_HANDLER_RESP_200 = {
     amount: 200
 }
 
+const BALANCE_HANDLER_RESP_200 = {
+    userId: 'player',
+    balance: 200
+}
+
 const ERROR_HANDLER_RESP = {
     statusCode: '400',
     body: {}
 }
 
-describe('Index: When query event is received', function() {
-    describe('When query controller returns successfully', function() {
+describe('Index: When transfer XP event is received', function() {
+    describe('When transfer handler returns successfully', function() {
         it('Returns HTTP 200', async function() {
-            const queryControllerMock = sinon.stub(queryHandler, "handle")
+            const transferHandlerMock = sinon.stub(transferHandler, "handle")
+                .returns()
+            const indexResp = await index.handler(EVENT_TRANSFER)
+            expect(indexResp.statusCode).to.be.equal('200')
+            transferHandlerMock.restore()
+        })
+    })
+    describe('When transfer handler throws an error', function() {
+        it('Returns error handler mapped response', async function() {
+            const transferHandlerMock = sinon.stub(transferHandler, "handle")
+                .throws()
+            const transferErrorHandlerMock = sinon.stub(transferErrorHandler, "handle")
+                .returns(ERROR_HANDLER_RESP)
+            const indexResp = await index.handler(EVENT_TRANSFER)
+            expect(indexResp.statusCode).to.be.deep.equal(ERROR_HANDLER_RESP.statusCode)
+            expect(JSON.parse(indexResp.body)).to.be.deep.equal(ERROR_HANDLER_RESP.body)
+            transferHandlerMock.restore()
+            transferErrorHandlerMock.restore()
+        })
+    })
+})
+
+describe('Index: When xp withdraw event is received', function() {
+    describe('When withdraw handler returns successfully', function() {
+        it('Returns HTTP 200', async function() {
+            const withdrawHandlerMock = sinon.stub(withdrawHandler, "handle")
+                .returns()
+            const indexResp = await index.handler(EVENT_WITHDRAW)
+            expect(indexResp.statusCode).to.be.equal('200')
+            withdrawHandlerMock.restore()
+        })
+    })
+    describe('When withdraw handler throws an error', function() {
+        it('Returns error handler mapped response', async function() {
+            const withdrawHandlerMock = sinon.stub(withdrawHandler, "handle")
+                .throws()
+            const withdrawErrorHandlerMock = sinon.stub(withdrawErrorHandler, "handle")
+                .returns(ERROR_HANDLER_RESP)
+            const indexResp = await index.handler(EVENT_WITHDRAW)
+            expect(indexResp.statusCode).to.be.deep.equal(ERROR_HANDLER_RESP.statusCode)
+            expect(JSON.parse(indexResp.body)).to.be.deep.equal(ERROR_HANDLER_RESP.body)
+            withdrawHandlerMock.restore()
+            withdrawErrorHandlerMock.restore()
+        })
+    })
+})
+
+describe('Index: When balance check event is received', function() {
+    describe('When balance handler returns successfully', function() {
+        it('Returns HTTP 200', async function() {
+            const balanceHandlerMock = sinon.stub(balanceHandler, "handle")
+                .returns(BALANCE_HANDLER_RESP_200)
+            const indexResp = await index.handler(EVENT_BALANCE)
+            expect(indexResp.statusCode).to.be.equal('200')
+            expect(JSON.parse(indexResp.body)).to.be.deep.equal(BALANCE_HANDLER_RESP_200)
+            balanceHandlerMock.restore()
+        })
+    })
+    describe('When balance handler throws an error', function() {
+        it('Returns error handler mapped response', async function() {
+            const balanceHandlerMock = sinon.stub(balanceHandler, "handle")
+                .throws()
+            const balanceErrorHandlerMock = sinon.stub(balanceErrorHandler, "handle")
+                .returns(ERROR_HANDLER_RESP)
+            const indexResp = await index.handler(EVENT_BALANCE)
+            expect(indexResp.statusCode).to.be.deep.equal(ERROR_HANDLER_RESP.statusCode)
+            expect(JSON.parse(indexResp.body)).to.be.deep.equal(ERROR_HANDLER_RESP.body)
+            balanceHandlerMock.restore()
+            balanceErrorHandlerMock.restore()
+        })
+    })
+})
+
+describe('Index: When deposit event is received', function() {
+    describe('When deposit handler returns successfully', function() {
+        it('Returns HTTP 200', async function() {
+            const depositHandlerMock = sinon.stub(depositHandler, "handle")
+                .returns()
+            const indexResp = await index.handler(EVENT_DEPOSIT)
+            expect(indexResp.statusCode).to.be.equal('200')
+            depositHandlerMock.restore()
+        })
+    })
+    describe('When deposit handler throws an error', function() {
+        it('Returns error handler mapped response', async function() {
+            const depositHandlerMock = sinon.stub(depositHandler, "handle")
+                .throws()
+            const errorHandlerMock = sinon.stub(depositErrorHandler, "handle")
+                .returns(ERROR_HANDLER_RESP)
+            const indexResp = await index.handler(EVENT_DEPOSIT)
+            expect(indexResp.statusCode).to.be.equal(ERROR_HANDLER_RESP.statusCode)
+            expect(JSON.parse(indexResp.body)).to.be.deep.equal(ERROR_HANDLER_RESP.body)
+            depositHandlerMock.restore()
+            errorHandlerMock.restore()
+        })
+    })
+})
+
+describe('Index: When query event is received', function() {
+    describe('When query handler returns successfully', function() {
+        it('Returns HTTP 200', async function() {
+            const queryHandlerMock = sinon.stub(queryHandler, "handle")
                 .returns(QUERY_HANDLER_RESP_200)
             const indexResp = await index.handler(EVENT_QUERY)
             expect(JSON.parse(indexResp.body)).to.be.deep.equal(QUERY_HANDLER_RESP_200)
             expect(indexResp.statusCode).to.be.equal('200')
-            queryControllerMock.restore()
+            queryHandlerMock.restore()
         })
     })
-    describe('When query controller throws an error', function() {
+    describe('When query handler throws an error', function() {
         it('Returns error handler mapped response', async function() {
-            const queryControllerMock = sinon.stub(queryHandler, "handle")
+            const queryHandlerMock = sinon.stub(queryHandler, "handle")
                 .throws()
             const errorHandlerMock = sinon.stub(queryErrorHandler, "handle")
                 .returns(ERROR_HANDLER_RESP)
             const indexResp = await index.handler(EVENT_QUERY)
             expect(indexResp.statusCode).to.be.equal(ERROR_HANDLER_RESP.statusCode)
             expect(JSON.parse(indexResp.body)).to.be.deep.equal(ERROR_HANDLER_RESP.body)
-            queryControllerMock.restore()
+            queryHandlerMock.restore()
             errorHandlerMock.restore()
         })
     })