+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 This project attempts to conform to [SemVer](https://semver.org/) guidelines whenever possible.
 
+## [v0.2.0]
+* [Implementing Deposit XP API](https://github.com/Ubunfu/mc-xp-bank/pull/1)
+* [Implementing XP Account Balance API](https://github.com/Ubunfu/mc-xp-bank/pull/2)
+* [Implementing Withdraw XP API](https://github.com/Ubunfu/mc-xp-bank/pull/3)
+* [Implementing Transfer XP API](https://github.com/Ubunfu/mc-xp-bank/pull/4)
+* [Fixing error in prod circle ci config](https://github.com/Ubunfu/mc-xp-bank/pull/5)
+
 ## [v0.1.0]
 * Implementing Query XP API